pyproject.toml
[project]
name = "content2subs"
version = "0.1.0"
description = ""
authors = [
    {name = "GiowGiow",email = "giofmartinelli@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.8,<3.11"
dependencies = [
    "openai-whisper (>=20240930,<20240931)",
    "ffmpeg-python (>=0.2.0,<0.3.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
